Components structure for sla view
